renamed files and add check to see if generated code exists

+++ TwitterCode/asynchronous_post_to_twitter_manager.py
@@ -1,5 +1,6 @@
 import asynchronous_twitter_code
 import math
+import os
 import requests
 from requests.structures import CaseInsensitiveDict
 from twython import Twython
@@ -7,6 +8,9 @@ import twython.exceptions
 
 
 def generate_asynchronous_code(values_map):
+    if os.path.exists('asynchronous_twitter_code.py'):
+        print('Generated asynchronous code already exists...')
+        return
     create_async_code_file = open('asynchronous_twitter_code.py', 'w')
     create_async_code_file.write('''import asyncio\nfrom post_multiple_to_twitter_obj import ManageFlowObj\n\n\n''')
     boiler_plate_code = '''    
@@ -74,7 +78,7 @@ class ManageMultipleTwitterPosts:
         self.create_map()
         generate_asynchronous_code(self.values_map)
         print('Beginning program...')
-        asynchronous_twitter_code.run(self.values_map)
+        # asynchronous_twitter_code.run(self.values_map)
 
     def validate_params(self):
         print('Beginning validation of Twitter Values File...')
